依照助教建議修改

+++ controller/postController.js
@@ -18,7 +18,7 @@ const postController = {
                         location: data.location,
                         type: data.type,
                         tags: data.tags,
-                        content: data.content,
+                        content: data.content.trim(),
                         image: data.image,
                         likes: data.likes
                     }
@@ -33,13 +33,19 @@ const postController = {
             errorHandle(res, message);
         };
     },
+    // DELETE 刪除單筆資料時，若未填寫 ID 路由為 "/posts/” 時，會刪除所有貼文。用 req.originalUrl 判斷路由是否為 "/posts/” ，是才執行
     async deleteAllPosts(req, res) {
-        await Post.deleteMany();
-        res.status(200).send({
-            "status": "success",
-            "message": "已刪除全部貼文"
-        });
-        res.end();
+        if (req.originalUrl !== "/posts/") {
+            await Post.deleteMany();
+            res.status(200).send({
+                "status": "success",
+                "message": "已刪除全部貼文"
+            })
+            res.end();
+        } else {
+            const message = '查無該筆貼文 id';
+            errorHandle(res, message);
+        }
     },
     async deletePosts(req, res) {
         const id = req.params.id;
@@ -48,7 +54,7 @@ const postController = {
         // console.log(idResult);
         if (idResult !== null) {
             successHandle(res, null);
-            
+
         } else {
             const message = '查無該筆貼文 id';
             errorHandle(res, message);
@@ -70,12 +76,13 @@ const postController = {
                         location: data.location,
                         type: data.type,
                         tags: data.tags,
-                        content: data.content,
+                        content: data.content.trim(),
                         image: data.image,
                         likes: data.likes
                     }
                 );
-                successHandle(res, data);
+                const newPost = await Post.findById(id);
+                successHandle(res, newPost);
             } else {
                 const message = '查無該筆貼文內容或 id 屬性';
                 errorHandle(res, message);
